Move builder traits to separate file
