+++ sdk/cosmos/src/prelude.rs
@@ -9,25 +9,8 @@ pub use crate::database::DatabaseName;
 pub use crate::document::Document;
 pub use crate::query::Query;
 pub use crate::responses::{QueryDocumentsResponse, QueryDocumentsResponseRaw, QueryResult};
+pub use crate::traits::*;
 pub use crate::{
-    AIMOption, AIMSupport, AllowTentativeWritesOption, AllowTentativeWritesSupport,
-    AttachmentClientRequired, AuthorizationToken, CollectionClientRequired, CollectionNameRequired,
-    CollectionNameSupport, CollectionRequired, CollectionSupport, ConsistencyLevel,
-    ConsistencyLevelOption, ConsistencyLevelSupport, CosmosClientRequired, CosmosError,
-    DatabaseClientRequired, DatabaseNameRequired, DatabaseNameSupport, DocumentIdRequired,
-    DocumentIdSupport, ExpirySecondsOption, ExpirySecondsSupport, IndexingDirective,
-    IndexingDirectiveOption, IndexingDirectiveSupport, IndexingPolicyRequired,
-    IndexingPolicySupport, IsUpsertOption, IsUpsertSupport, MaxItemCountOption,
-    MaxItemCountSupport, MediaRequired, MediaSupport, Offer, OfferRequired, OfferSupport,
-    ParallelizeCrossPartitionQueryOption, ParallelizeCrossPartitionQuerySupport, ParametersOption,
-    ParametersSupport, PartitionKeyOption, PartitionKeyRequired, PartitionKeySupport,
-    PartitionKeys, PartitionKeysOption, PartitionKeysRequired, PartitionKeysSupport,
-    PartitionRangeIdOption, PartitionRangeIdSupport, PermissionClientRequired,
-    QueryCrossPartitionOption, QueryCrossPartitionSupport, QueryRequired, QuerySupport,
-    StoredProcedureBodyRequired, StoredProcedureBodySupport, StoredProcedureClientRequired,
-    StoredProcedureNameRequired, StoredProcedureNameSupport, TriggerBodyRequired,
-    TriggerBodySupport, TriggerClientRequired, TriggerOperationRequired, TriggerOperationSupport,
-    TriggerTypeRequired, TriggerTypeSupport, UserClientRequired, UserDefinedFunctionBodyRequired,
-    UserDefinedFunctionBodySupport, UserDefinedFunctionClientRequired, UserName, UserNameRequired,
-    UserNameSupport,
+    AuthorizationToken, ConsistencyLevel, CosmosError, IndexingDirective, Offer, PartitionKeys,
+    UserName,
 };
