+++ sdk/cosmos/src/requests/delete_document_builder.rs
@@ -1,6 +1,5 @@
 use crate::prelude::*;
 use crate::responses::DeleteDocumentResponse;
-use crate::DocumentClientRequired;
 use azure_core::modify_conditions::IfMatchCondition;
 use azure_core::prelude::*;
 use azure_core::{IfMatchConditionOption, IfMatchConditionSupport};
