+++ sdk/cosmos/src/requests/get_collection_builder.rs
@@ -1,6 +1,5 @@
 use crate::prelude::*;
 use crate::responses::GetCollectionResponse;
-use crate::CollectionClientRequired;
 use azure_core::prelude::*;
 use http::StatusCode;
 use std::convert::TryInto;
