+++ sdk/cosmos/src/requests/list_attachments_builder.rs
@@ -1,6 +1,5 @@
 use crate::prelude::*;
 use crate::responses::ListAttachmentsResponse;
-use crate::DocumentClientRequired;
 use crate::{DocumentClient, ResourceType};
 use azure_core::prelude::*;
 use futures::stream::{unfold, Stream};
