+++ sdk/cosmos/src/requests/list_collections_builder.rs
@@ -1,6 +1,6 @@
 use crate::prelude::*;
 use crate::responses::ListCollectionsResponse;
-use crate::{DatabaseClientRequired, ResourceType};
+use crate::ResourceType;
 use azure_core::prelude::*;
 use futures::stream::{unfold, Stream};
 use http::StatusCode;
